api/requirements.txt
fastapi
uvicorn
pydantic
elasticsearch
google-cloud-documentai
google-cloud-storage
python-multipart
python-dotenv
google-cloud-aiplatform
langdetect
